[hir2mpl]Internal commit msg:     [IPA] code clean.     [hir2mpl] clean up pclint in common, part 2     [hir2mpl] clean up pclint alarm in ast_input [mapleall]Internal commit msg:     [mplcg]Enable ElimSpecificAnd     [IPA] code clean.     [mplcg] prepeephole LongIntCompareWithZPattern use AArch64CGPeepHole     !1053 Added "--calleeregs-placement" opt. Default is off for now.     !1033 [mapleall] Fix bugs related to iterators invalidation

+++ src/hir2mpl/common/include/feir_stmt.h
@@ -302,7 +302,7 @@ class FEIRStmt : public FELinkListNode {
 
   void SetMIRStmtSrcPos(std::list<StmtNode*> &stmts) const {
     if (FEOptions::GetInstance().IsDumpLOC() && !stmts.empty()) {
-      (*stmts.begin())->GetSrcPos().SetFileNum(srcFileIndex);
+      (*stmts.begin())->GetSrcPos().SetFileNum(static_cast<uint16>(srcFileIndex));
       (*stmts.begin())->GetSrcPos().SetLineNum(srcFileLineNum);
     }
   }
@@ -680,7 +680,7 @@ class FEIRExprAddrofConstArray : public FEIRExpr {
   ~FEIRExprAddrofConstArray() = default;
 
   uint32 GetStringLiteralSize() const {
-    return array.size();
+    return static_cast<uint32>(array.size());
   }
 
   const MIRType *GetElemType() const {
